// src/app/business-logic/model/projects/projectsGetHyperparamValuesResponse.ts
/**
 * projects
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface ProjectsGetHyperparamValuesResponse {
    /**
     * Total number of distinct parameter values
     */
    total?: number;
    /**
     * The list of the unique values for the parameter
     */
    values?: Array<string>;
}