// src/app/business-logic/model/tasks/tasksResetManyResponse.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { TasksResetManyResponseFailed } from '././tasksResetManyResponseFailed';
import { TasksResetManyResponseSucceeded } from '././tasksResetManyResponseSucceeded';


export interface TasksResetManyResponse {
    succeeded?: Array<TasksResetManyResponseSucceeded>;
    failed?: Array<TasksResetManyResponseFailed>;
}